2.0
IMPORTANT! Only update if you already updated Yoast SEO plugin to 13.0 or above! This version will work only with the new version. If you didn't update the Yoast SEO plugin yet, you can keep the 1.1 version of this plugin also, it is working without any problem.

It is always recommended to use the latest versions of plugins, so please do update Yoast SEO to 13.0 or above and do the update also for my plugin!

FIX

- Yoast SEO 13.0 changed the option for XML sitemaps. (Props: @iarovuo)

TWEAK

- Minor changes in title and description.
- Tested with WordPress 5.3 version.
- Updated to use a new way to deploy it to wp.org repo.